+++ CHANGELOG.md
@@ -6,9 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added simple retry policy configuration to superjson schema
+- Added `openTime` field to circuit breaker configuration in superjson schema
+
 ### Changed
 - removed `typescript-is` and replaced it with `ajv` validation
 - `isProfileDocumentNode` and `isMapDocumentNode` do full schema validation
+- Cleaned up schema types by removing now obsolete `| undefined`.
 
 ## [1.1.0] - 2022-01-19
 ### Added